+++ modules/core/src/test/scala/moe/brianhsu/live2d/enitiy/avatar/effect/impl/PhysicsFeature.scala
@@ -128,7 +128,7 @@ class PhysicsFeature extends AnyFeatureSpec with GivenWhenThen with Matchers wit
   private def createStubbedModel(logData: LogData): Live2DModel = {
     val model: Live2DModel = stub[Live2DModel]
     val mockedParameters = logData.parameters.map { case (id, parameter) =>
-      id -> new JavaVMParameter(id, ParameterType.Normal, parameter.min, parameter.max, parameter.default, parameter.current)
+      id -> new JavaVMParameter(id, ParameterType.Normal, parameter.min, parameter.max, parameter.default, Nil, parameter.current)
     }
 
     (() => model.parameters).when().returns(mockedParameters)